pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "my-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.12",
    "langchain-openrouter>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "psycopg2-binary>=2.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
